+++ src/components/footer.js
@@ -3,73 +3,51 @@ import GitHubIcon from '@mui/icons-material/GitHub';
 import InfoIcon from '@mui/icons-material/Info';
 import MailIcon from '@mui/icons-material/Mail';
 import LinkedInIcon from '@mui/icons-material/LinkedIn';
-import PhoneIcon from '@mui/icons-material/Phone';
-import { Box, Container, Typography, Grid, Divider, Link } from "@mui/material";
+import { Box, Link, Grid } from "@mui/material";
 
 const Footer = () => (
-  <Box sx={{ bgcolor: 'black', p: 6, mt: 'auto', color: 'white' }}>
-    <Container maxWidth="sm">
-      <Grid container spacing={3} justifyContent="center">
-        <Grid item>
-          <Typography variant="body1">
-            <Link
-              href="/about"
-              underline="hover"
-              sx={{ color: 'white', '&:hover': { color: 'white' }, '&:active': { color: 'grey' } }}
-            >
-              <InfoIcon sx={{ verticalAlign: 'middle' }} /> About
-            </Link>
-          </Typography>
-        </Grid>
-        <Grid item>
-          <Divider orientation="vertical" flexItem />
-        </Grid>
-        <Grid item>
-          <Typography variant="body1">
-            <Link
-              href="https://github.com/your-github-username"
-              underline="hover"
-              sx={{ color: 'white', '&:hover': { color: 'white' }, '&:active': { color: 'grey' } }}
-            >
-              <GitHubIcon sx={{ verticalAlign: 'middle' }} /> GitHub
-            </Link>
-          </Typography>
-        </Grid>
-        <Grid item>
-          <Divider orientation="vertical" flexItem />
-        </Grid>
-        <Grid item>
-          <Typography variant="body1">Contact</Typography>
-          <Typography variant="body2">
-            <Link
-              href="mailto:your-email@example.com"
-              underline="hover"
-              sx={{ color: 'white', '&:hover': { color: 'white' }, '&:active': { color: 'grey' } }}
-            >
-              <MailIcon sx={{ verticalAlign: 'middle' }} /> Email
-            </Link>
-          </Typography>
-          <Typography variant="body2">
-            <Link
-              href="tel:123-456-7890"
-              underline="hover"
-              sx={{ color: 'white', '&:hover': { color: 'white' }, '&:active': { color: 'grey' } }}
-            >
-              <PhoneIcon sx={{ verticalAlign: 'middle' }} /> Phone
-            </Link>
-          </Typography>
-          <Typography variant="body2">
-            <Link
-              _href="https://www.linkedin.com/in/your-linkedin-username"
-              underline="hover"
-              sx={{ color: 'white', '&:hover': { color: 'white' }, '&:active': { color: 'grey' } }}
-            >
-              <LinkedInIcon sx={{ verticalAlign: 'middle' }} /> LinkedIn
-            </Link>
-          </Typography>
-        </Grid>
+  <Box sx={{ bgcolor: 'black', p: 2, mt: 'auto', color: 'white' }}>
+    <Grid container spacing={2} justifyContent="center">
+      <Grid item>
+        <Link
+          href="/about"
+          underline="hover"
+          sx={{ color: 'white', '&:hover': { color: 'white' }, '&:active': { color: 'grey' } }}
+        >
+          <InfoIcon sx={{ verticalAlign: 'middle' }} />
+        </Link>
       </Grid>
-    </Container>
+      <Grid item>
+        <Link
+          href="https://github.com/antch57/jam-statz-frontend" // replace with your GitHub username
+          underline="hover"
+          target="_blank"
+          sx={{ color: 'white', '&:hover': { color: 'white' }, '&:active': { color: 'grey' } }}
+        >
+          <GitHubIcon sx={{ verticalAlign: 'middle' }} />
+        </Link>
+      </Grid>
+      <Grid item>
+        <Link
+          href="mailto:actual-email@example.com" // replace with your email
+          underline="hover"
+          target="_blank"
+          sx={{ color: 'white', '&:hover': { color: 'white' }, '&:active': { color: 'grey' } }}
+        >
+          <MailIcon sx={{ verticalAlign: 'middle' }} />
+        </Link>
+      </Grid>
+      <Grid item>
+        <Link
+          href="https://www.linkedin.com/in/anthony-chiarello-271a46132/" // replace with your LinkedIn username
+          underline="hover"
+          target="_blank"
+          sx={{ color: 'white', '&:hover': { color: 'white' }, '&:active': { color: 'grey' } }}
+        >
+          <LinkedInIcon sx={{ verticalAlign: 'middle' }} />
+        </Link>
+      </Grid>
+    </Grid>
   </Box>
 );
