+++ src/pages/index.js
@@ -1,13 +1,15 @@
 import React from "react"
 import Layout from '../components/layout';
-
+import { Box, Typography, Container } from '@mui/material';
 
 const HomePage = () => (
   <Layout>
-    <div style={{ margin: '3rem auto', maxWidth: 600 }}>
-      <h1 style={{ textAlign: 'center', color: 'purple' }}>Welcome to My Website!</h1>
-      <p style={{ textAlign: 'center' }}>im trying so gaawwlll.</p>
-    </div>
+    <Container>
+      <Box sx={{ margin: '3rem auto', maxWidth: 600, textAlign: 'center' }}>
+        <Typography variant="h1" color="purple">Welcome to My Website!</Typography>
+        <Typography>im trying so gaawwlll.</Typography>
+      </Box>
+    </Container>
   </Layout>
 )
